TWITTER_FUNCTION/requirements.txt
# Function dependencies, for example:
# package>=version
tweepy
pandas
google.cloud.storage
gcloud
google.cloud.bigquery